requirements.txt
cassandra-driver
pandas
matplotlib
seaborn
ipywidgets
requests
statsmodels
pyspark
plotly
streamsync
numpy
